project2/nothotdog/requirements.txt
absl-py==0.11.0
asgiref==3.3.1
astunparse==1.6.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
Django==3.1.5
django-appconf==1.0.4
django-compressor==2.4
django-debug-toolbar==3.2
djangorestframework==3.12.2
flake8==3.8.4
flatbuffers==1.12
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
joblib==1.0.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.4
mccabe==0.6.1
nameko==2.13.0
networkx==2.5
numpy==1.19.5
oauthlib==3.1.0
opencv-python-headless==4.5.1.48
opt-einsum==3.3.0
pika==1.1.0
Pillow==8.1.0
pip==20.0.2
protobuf==3.14.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
PyWavelets==1.1.1
PyYAML==5.4.1
rcssmin==1.0.6
requests==2.25.1
requests-oauthlib==1.3.0
rjsmin==1.1.0
rsa==4.7
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.0
setuptools==44.0.0
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2021.2.1
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
wheel==0.36.2
wrapt==1.12.1
